+++ tests/event_test.php
@@ -30,7 +30,6 @@ use mod_examcheck\local\steps;
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 final class event_test extends \advanced_testcase {
-
     /**
      * Marking fires user_marked; unmarking fires user_unmarked.
      */